Changed an incorrect reference to:
$SOSROOT/modules/test_modules/surge/

to the correct reference:

$SOSROOT/modules/demos/surge/

Things to cosider: As send to the list, there are places where modules/sensordrivers/mts310/photosensor/ or modules/sensordrivers/micasb/photosensor/ are referenced. In my copy these directories are empty, and their last commit was r72. modules/sensordrivers/mts310/photo_temp/ has code in it and was last touched around r100. It seems as if the tutorial should be updated to reflect this, but because it is a non-cosmetic change, I'm leaving it for others to decide.

-Peter Mawhorter