Added new demo program, Ping Pong
Ping Pong is a small little program to test the two way radio communication of two nodes. Node 1 sends out a regular ping message, and node 0 responds with a Pong. Bot nodes report to the UART the sequence number of the current ping/pong message they just sent. This helps to check for message losses etc.